requirements.txt
streamlit==1.32.0
opencv-python-headless==4.9.0.80
pdf2image==1.17.0
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
fpdf==1.7.2